# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.3.0
  description: SOC OS - SaaS Security Platform API (V30/V50)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agent/observe:
    post:
      operationId: runObserverAgent
      tags: [agent]
      summary: Run a policy-gated observer-only analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObserverAgentInput"
      responses:
        "200":
          description: Observer analysis completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObserverAgentRun"
        "400":
          description: Invalid observer request
        "403":
          description: Policy blocked the requested action
...
  /correlations:
    get:
      operationId: listCorrelations
      tags: [correlations]
      summary: List detected attack chain correlations
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: severity
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of correlations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Correlation"
  /correlations/{id}/resolve:
    patch:
      operationId: resolveCorrelation
      tags: [correlations]
      summary: Mark a correlation as resolved
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved correlation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Correlation"
        "404":
          description: Not found

  # ─── VIRUS INTELLIGENCE DATABASE ───────────────────────────────────────────
...
components:
  schemas:
...
    ObserverAgentInput:
      type: object
      properties:
        taskType:
          type: string
          enum: [analyze_event, analyze_system, inspect_memory]
        instruction:
          type: string
          minLength: 1
          maxLength: 2000
        event:
          type: string
          maxLength: 10000
        tenantId:
          type: integer
          minimum: 1
      required: [taskType, instruction]
    ObserverAgentRun:
      type: object
      properties:
        runId:
          type: string
          format: uuid
        status:
          type: string
          enum: [blocked, completed, failed]
        input:
          $ref: "#/components/schemas/ObserverAgentInput"
        policy:
          type: object
          additionalProperties: true
        plan:
          type: array
          items:
            type: object
            additionalProperties: true
        results:
          type: array
          items:
            type: object
            additionalProperties: true
        reflection:
          type: object
          additionalProperties: true
        productionChanged:
          type: boolean
          enum: [false]
      required: [runId, status, input, policy, plan, results, reflection, productionChanged]
...
    Correlation:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: ["integer", "null"]
        threatType:
          type: string
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        confidence:
          type: number
        summary:
          type: string
        eventIds:
          type: string
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, threatType, severity, confidence, summary, eventIds, createdAt]
...